add templates for Create/List Views and fixing some bugs

+++ Domain.Api/Pages/Admin/User/DeleteUser.cshtml.cs
@@ -1,5 +1,4 @@
 using Domain.Services.Contracts.Services;
-using Domain.Services.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
@@ -14,16 +13,19 @@ namespace Domain.Api.Pages.Admin.User
             this.userService = userService;
         }
 
-        public IActionResult OnGet(Guid? id)
+        public async Task<IActionResult> OnGetAsync(Guid? id)
         {
-            var user = userService.GetByIdAsync(id);
+            var user = await userService.GetByIdAsync(id);
 
             if (user != null)
             {
-                userService.DeleteAsync(id);
+                if (!(user.Role == Generics.Enums.UserRole.Admin && await userService.GetCountAsync(user.Role) <= 1))
+                {
+                    await userService.DeleteAsync(id);
+                }
             }
 
-            return Redirect("/Admin/Users/UsersList");
+            return Redirect("/Admin/User/UserList");
         }
     }
 }
